refactor of test suite
