+++ tests/e2e/setup.go
@@ -1,62 +1,27 @@
 package e2e
 
 import (
-	"path/filepath"
 	"testing"
 
-	"cosmossdk.io/math"
 	"github.com/CosmWasm/wasmd/x/wasm/ibctesting"
 	wasmkeeper "github.com/CosmWasm/wasmd/x/wasm/keeper"
 	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
 	"github.com/cometbft/cometbft/types"
-	types2 "github.com/cosmos/ibc-go/v8/modules/core/04-channel/types"
-	ibctesting2 "github.com/cosmos/ibc-go/v8/testing"
 	"github.com/stretchr/testify/require"
 
 	"github.com/babylonchain/babylon-sdk/demo/app"
+	bstypes "github.com/babylonchain/babylon/x/btcstaking/types"
+	zctypes "github.com/babylonchain/babylon/x/zoneconcierge/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 )
 
-var (
-	wasmContractPath    string
-	wasmContractGZipped bool
-)
-
-func buildPathToWasm(fileName string) string {
-	if wasmContractGZipped {
-		fileName += ".gz"
-	}
-	return filepath.Join(wasmContractPath, fileName)
-}
-
-type Example struct {
-	t                *testing.T
-	Coordinator      *ibctesting.Coordinator
-	ConsumerChain    *ibctesting.TestChain
-	ProviderChain    *ibctesting.TestChain
-	ConsumerApp      *app.ConsumerApp
-	IbcPath          *ibctesting.Path
-	ProviderDenom    string
-	ConsumerDenom    string
-	MyProvChainActor string
-}
-
-func NewExample(t *testing.T) Example {
-	coord := NewIBCCoordinator(t)
-	provChain := coord.GetChain(ibctesting2.GetChainID(1))
-	consChain := coord.GetChain(ibctesting2.GetChainID(2))
-	return Example{
-		t:                t,
-		Coordinator:      coord,
-		ConsumerChain:    consChain,
-		ProviderChain:    provChain,
-		ConsumerApp:      consChain.App.(*app.ConsumerApp),
-		IbcPath:          ibctesting.NewPath(consChain, provChain),
-		ProviderDenom:    sdk.DefaultBondDenom,
-		ConsumerDenom:    sdk.DefaultBondDenom,
-		MyProvChainActor: provChain.SenderAccount.GetAddress().String(),
+func NewBTCStakingPacketData(packet *bstypes.BTCStakingIBCPacket) *zctypes.ZoneconciergePacketData {
+	return &zctypes.ZoneconciergePacketData{
+		Packet: &zctypes.ZoneconciergePacketData_BtcStaking{
+			BtcStaking: packet,
+		},
 	}
 }
 
@@ -102,40 +67,4 @@ func InstantiateContract(t *testing.T, chain *ibctesting.TestChain, codeID uint6
 	require.Len(t, contractAddrs, 1)
 
 	return contractAddrs[0]
-}
-
-func (x *Example) SetupBabylonIntegration() (*TestConsumerClient, *ConsumerContract, *TestProviderClient) {
-	x.Coordinator.SetupConnections(x.IbcPath)
-
-	// setup contracts on consumer
-	consumerCli := NewConsumerClient(x.t, x.ConsumerChain)
-	consumerContracts, err := consumerCli.BootstrapContracts()
-	require.NoError(x.t, err)
-	consumerPortID := wasmkeeper.PortIDForContract(consumerContracts.Babylon)
-
-	// add some fees so that we can distribute something
-	x.ConsumerChain.DefaultMsgFees = sdk.NewCoins(sdk.NewCoin(x.ConsumerDenom, math.NewInt(1_000_000)))
-
-	providerCli := NewProviderClient(x.t, x.ProviderChain)
-
-	return consumerCli, consumerContracts, providerCli
-
-	// TODO: fix IBC channel below
-
-	// setup ibc path
-	x.IbcPath.EndpointB.ChannelConfig = &ibctesting2.ChannelConfig{
-		PortID: "zoneconcierge", // TODO: replace this chain/port with Babylon
-		Order:  types2.ORDERED,
-	}
-	x.IbcPath.EndpointA.ChannelConfig = &ibctesting2.ChannelConfig{
-		PortID: consumerPortID,
-		Order:  types2.ORDERED,
-	}
-	x.Coordinator.CreateChannels(x.IbcPath)
-
-	// when ibc package is relayed
-	require.NotEmpty(x.t, x.ConsumerChain.PendingSendPackets)
-	require.NoError(x.t, x.Coordinator.RelayAndAckPendingPackets(x.IbcPath))
-
-	return consumerCli, consumerContracts, providerCli
 }
